CODE REFACTOR
- Replaced private methods in MetabaseServer with imported handler functions for better modularity and maintainability.
